+++ app/Http/Controllers/LessonController.php
@@ -18,11 +18,19 @@ class LessonController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create($discipline_id)
+    public function create($discipline_id, Request $request)
     {
-        $discipline = Discipline::findOrFail($discipline_id);
-        $groups = Group::all(); 
-        return view('teacher.disciplines.lessons.create', compact('discipline', 'groups'));
+        $user = Auth::user();
+        // проверка прав пользователя
+        if ($request->user()->can('create', User::class)) {
+            $discipline = Discipline::findOrFail($discipline_id);
+            $groups = Group::all(); 
+            return view('teacher.disciplines.lessons.create', compact('discipline', 'groups'));
+        } else {
+            // запрет действия с выводом сообщения об ошибке доступа
+            return redirect()->route('home')
+                ->withErrors(['msg' => 'Ошибка доступа']);
+        }
     }
 
     /**
@@ -33,12 +41,18 @@ class LessonController extends Controller
      */
     public function store(LessonRequest $request, $discipline_id)
     {
-        $discipline = Discipline::findOrFail($discipline_id);
-        $request['discipline_id'] = $discipline->id;
-        $group = Group::findOrFail($request->group_id);
-        $lesson = Lesson::create($request->all());
-        $lesson->students()->attach($group->students);
-        return redirect()->route('teacher.lessons.show', [$discipline, $lesson]);
+        if ($request->user()->can('create', User::class)) {
+            $discipline = Discipline::findOrFail($discipline_id);
+            $request['discipline_id'] = $discipline->id;
+            $group = Group::findOrFail($request->group_id);
+            $lesson = Lesson::create($request->all());
+            $lesson->students()->attach($group->students);
+            return redirect()->route('teacher.lessons.show', [$discipline, $lesson]);
+        } else {
+            // запрет действия с выводом сообщения об ошибке доступа
+            return redirect()->route('home')
+                ->withErrors(['msg' => 'Ошибка доступа']);
+        }
     }
 
     /**
@@ -47,11 +61,19 @@ class LessonController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($discipline_id, $id)
+    public function show($discipline_id, Request $request, $id)
     {
-        $discipline = Discipline::findOrFail($discipline_id);
-        $lesson = Lesson::findOrFail($id);
-        return view('teacher.disciplines.lessons.show', compact('discipline', 'lesson'));
+        $user = User::findOrFail($id);
+        // проверка прав пользователя
+        if ($request->user()->can('viewAny', $user)) {
+            $discipline = Discipline::findOrFail($discipline_id);
+            $lesson = Lesson::findOrFail($id);
+            return view('teacher.disciplines.lessons.show', compact('discipline', 'lesson'));
+        } else {
+            // запрет действия с выводом сообщения об ошибке доступа
+            return redirect()->route('home')
+                ->withErrors(['msg' => 'Ошибка доступа']);
+        }
     }
 
     /**
@@ -60,12 +82,19 @@ class LessonController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($discipline_id, $id)
+    public function edit($discipline_id, Request $request, $id)
     {
-        $discipline = Discipline::findOrFail($discipline_id);
-        $lesson = Lesson::findOrFail($id);
-        $groups = Group::all(); 
-        return view('teacher.disciplines.lessons.edit', compact('discipline', 'lesson', 'groups'));
+        $user = User::findOrFail($id);
+        if ($request->user()->can('update', $user)) {
+            $discipline = Discipline::findOrFail($discipline_id);
+            $lesson = Lesson::findOrFail($id);
+            $groups = Group::all(); 
+            return view('teacher.disciplines.lessons.edit', compact('discipline', 'lesson', 'groups'));
+        } else {
+            // запрет действия с выводом сообщения об ошибке доступа
+            return redirect()->route('home')
+                ->withErrors(['msg' => 'Ошибка доступа']);
+        }
     }
 
     /**
@@ -77,10 +106,18 @@ class LessonController extends Controller
      */
     public function update(LessonRequest $request, $discipline_id, $id)
     {
-        $discipline = Discipline::findOrFail($discipline_id);
-        $lesson = Lesson::findOrFail($id);
-        $lesson->update($request->all());
-        return redirect()->route('teacher.disciplines.show', $discipline);
+        $user = User::findOrFail($id);
+        // проверка прав пользователя
+        if ($request->user()->can('update', $user)) {
+            $discipline = Discipline::findOrFail($discipline_id);
+            $lesson = Lesson::findOrFail($id);
+            $lesson->update($request->all());
+            return redirect()->route('teacher.disciplines.show', $discipline);
+        } else {
+            // запрет действия с выводом сообщения об ошибке доступа
+            return redirect()->route('home')
+                ->withErrors(['msg' => 'Ошибка доступа']);
+        }
     }
 
     /**
@@ -91,9 +128,17 @@ class LessonController extends Controller
      */
     public function destroy(Request $request, $discipline_id, $id)
     {
-        $discipline = Discipline::findOrFail($discipline_id);
-        $lesson = Lesson::findOrFail($id);
-        $lesson->delete();
-        return redirect()->route('teacher.disciplines.show', $discipline);
+        $user = User::findOrFail($id);
+        // проверка прав пользователя
+        if ($request->user()->can('delete', $user)) {
+            $discipline = Discipline::findOrFail($discipline_id);
+            $lesson = Lesson::findOrFail($id);
+            $lesson->delete();
+            return redirect()->route('teacher.disciplines.show', $discipline);
+        } else {
+            // запрет действия с выводом сообщения об ошибке доступа
+            return redirect()->route('home')
+                ->withErrors(['msg' => 'Ошибка доступа']);
+        }
     }
 }
